+++ src/core/auth/user.model.ts
