+++ src/core/flag-manager/model/category.model.ts
