+++ src/core/flag-manager/model/flag.model.ts
