+++ src/core/flag-manager/service/index.ts
@@ -2,3 +2,4 @@ export * from './audit.service';
 export * from './category.service';
 export * from './flag.service';
 export * from './flag-evaluation.service';
+export * from './flag-ttl.service';
