+++ src/infrastructure/config/container.ts
@@ -4,16 +4,11 @@ import { persistenceModule } from '../persistence';
 import { storageModule, OnInit, OnDestroy } from '../storage';
 import { ConfigService } from './config';
 import { TYPES } from './types';
-import {
-  AuditService,
-  CategoryService,
-  FeatureFlagService,
-  FlagEvaluationService,
-} from '../../core/service';
-import { FlagTTLService } from '../../core/service/flag-ttl.service';
+
 import { FlagHttpController } from '../delivery/api/controllers/v1/flag.http.controller';
 import { AuditHttpController } from '../delivery/api/controllers/v1/audit.http.controller';
 import { CategoryHttpController } from '../delivery/api/controllers/v1/category.http.controller';
+import { flagManagerContainer } from '../../core/flag-manager';
 
 export async function createContainer(): Promise<Container> {
   const container = new Container();
@@ -23,19 +18,7 @@ export async function createContainer(): Promise<Container> {
 
   container.load(storageModule);
   container.load(persistenceModule);
-
-  container.bind<AuditService>(TYPES.AuditService).to(AuditService).inSingletonScope();
-  container.bind<CategoryService>(TYPES.CategoryService).to(CategoryService).inSingletonScope();
-  container
-    .bind<FeatureFlagService>(TYPES.FeatureFlagService)
-    .to(FeatureFlagService)
-    .inSingletonScope();
-  container
-    .bind<FlagEvaluationService>(TYPES.FlagEvaluationService)
-    .to(FlagEvaluationService)
-    .inSingletonScope();
-
-  container.bind<FlagTTLService>(TYPES.FlagTTLService).to(FlagTTLService).inSingletonScope();
+  container.load(flagManagerContainer);
 
   container
     .bind<AuditHttpController>(TYPES.AuditHttpController)
