+++ src/shared/kernel/global.types.ts
