+++ tsconfig.json
@@ -15,6 +15,6 @@
     "emitDecoratorMetadata": true,
     "resolveJsonModule": true
   },
-  "include": ["src/**/*", "scripts/seed.js", "src/shared/kernel/global.ts"],
+  "include": ["src/**/*", "scripts/seed.js"],
   "exclude": ["node_modules", "**/*.test.ts"]
 }
